Moved wal to storage/block
