+++ pkg/storage/block/complete_block.go
@@ -1,4 +1,4 @@
-package wal
+package block
 
 import (
 	"bytes"
