+++ pkg/storage/block/config.go
@@ -1,4 +1,4 @@
-package wal
+package block
 
 type Config struct {
 	Filepath string `yaml:"path"`
