+++ pkg/storage/block/head_block.go
@@ -1,4 +1,4 @@
-package wal
+package block
 
 import (
 	"bytes"
