+++ pkg/storage/trace_backend/local/local.go
@@ -62,7 +62,7 @@ func (rw *readerWriter) Write(_ context.Context, blockID uuid.UUID, tenantID str
 	}
 
 	// copy traces file.
-	//  todo:  consider having the storage backend responsible for removing the WAL.  in this case we could just
+	//  todo:  consider having the storage backend responsible for removing the block.  in this case we could just
 	//   do a rename here which would be way faster.
 	src, err := os.Open(tracesFilePath)
 	if err != nil {
